refactor(mantaray): split lib.rs into constants, entry, and manifest modules
Extract Entry, Manifest/ManifestIter, and internal constants into
dedicated files, making lib.rs a thin re-export hub. Move tests into
their respective modules and replace the crate-local keccak256 wrapper
with direct use of alloy_primitives.
